allow editing and deleting only for authors

+++ src/main/java/org/arjunaoverdrive/app/services/WordServiceImpl.java
@@ -4,6 +4,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.arjunaoverdrive.app.dao.WordRepository;
 import org.arjunaoverdrive.app.model.Word;
 import org.arjunaoverdrive.app.model.WordSet;
+import org.arjunaoverdrive.app.services.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,15 +15,21 @@ import java.util.List;
 public class WordServiceImpl implements WordService {
 
     private final WordRepository wordRepository;
+    private final UserService userService;
 
     @Autowired
-    public WordServiceImpl(WordRepository wordRepository) {
+    public WordServiceImpl(WordRepository wordRepository, UserService userService) {
         this.wordRepository = wordRepository;
+        this.userService = userService;
     }
 
     @Override
     public void deleteWord(Long id) {
         Word word = this.wordRepository.findById(id).get();
+        WordSet ws = word.getWordSet();
+        if(!ws.getCreatedBy().equals(userService.getUserFromSecurityContext())){
+            throw new RuntimeException("Cannot edit another user's set");
+        }
         this.wordRepository.delete(word);
         log.info("delete word " + id);
     }
